requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.7.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
databases==0.6.2
fastapi==0.85.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.20.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1
